Extract sse and http components from Bitloops Class
