--- src/Subscriber/index.ts
@@ -1,3 +1,0 @@
-// import
-// all sse things here
-// class Subscriber {}
