+++ src/definitions.ts
@@ -85,24 +85,6 @@ export type BitloopsConfig = {
   auth?: AuthenticationOptionsType;
 };
 
-// export type AxiosHandlerOutcome = [AxiosResponse, null] | [AxiosResponse | null, AxiosError] | [null, unknown];
-export type AxiosHandlerOutcome = AxiosDataResponse | AxiosErrorResponse | AxiosUnexpectedResponse;
-
-type AxiosDataResponse = {
-  data: AxiosResponse;
-  error: null;
-};
-
-type AxiosErrorResponse = {
-  data: AxiosResponse | undefined;
-  error: AxiosError;
-};
-
-type AxiosUnexpectedResponse = {
-  data: null;
-  error: unknown;
-};
-
 export const enum StorageKeys {
   USER_DATA = 'bitloops.auth.userData',
   BITLOOPS_CONFIG = 'bitloops.config',
